feat: Add word_dict.json with new words and their synonyms
The code changes include adding a new file, word_dict.json, which contains a dictionary of words and their synonyms. This file is used for word substitution in the sentiment analysis module. The purpose of this change is to improve the accuracy of sentiment analysis by considering a wider range of words and their meanings.

Recent user commits:
- sentiment
- sentiment word sentence index
- finish the data_process
- add some py file

Recent repository commits:
- sentiment
- sentiment word sentence index
- finish the data_process
- add some py file
- Initial commit